+++ src/lib/Components/Consignments/__tests__/index-tests.tsx
@@ -4,7 +4,7 @@ import * as renderer from "react-test-renderer"
 
 import Consignments from "../"
 
-it("Sets up the right view hierarchy", () => {
+it("Sets up the right navigator view hierarchy", () => {
   const tree = renderer.create(<Consignments />).toJSON()
   expect(tree).toMatchSnapshot()
 })