+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Fixes for the camera roll order in consignments - orta
 * Adds the new photo and selects it when taking a photo - orta
 * Adds a confirmation screen to the end of the consignment submission process - maxim/orta
+* Re-align headers in all Home tabs - alloy
 
 ### 1.4.0-beta.12