+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### 1.4.0-beta.13
 
+* Message timestamps are more human-readable with 'ago' - sarah
 * Fixes for the camera roll order in consignments - orta
 * Adds the new photo and selects it when taking a photo - orta
 * Adds a confirmation screen to the end of the consignment submission process - maxim/orta