+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Fix a bug that would not let an infinite scroll of artworks stop requesting more artworks - orta
 * More pixel pushing in the consignments welcome screen - alloy
 * More pixel pushing in the conversations zero state screen - alloy
+* Add saves and follows analytics - maxim
 
 ### 1.4.0-beta.13