+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ### Master
 
+### 1.4.0-rc.0
+
 * Make sure Home/Fairs rail does not clip any icons at the end of the list - alloy
 * Re-align headers in all Home tabs - alloy
 * Fix crash that would occur when tapping artwork component on inquiry view - alloy